Fixed double render issue
