+++ packages/core/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@framejs/core",
-  "version": "1.0.0-rc.1",
+  "version": "1.0.0-rc.3",
   "description": "A ultra light library for writing custom elements",
   "main": "dist/index.js",
   "files": [
