+++ packages/renderer-lit-html/package.json
@@ -1,12 +1,11 @@
 {
   "name": "@framejs/renderer-lit-html",
-  "version": "1.0.0-rc.1",
+  "version": "1.0.0-rc.3",
   "description": "A Framejs renderer for lit-html",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/renderer-lit-html.js",
+  "types": "dist/renderer-lit-html.d.ts",
   "scripts": {
-    "build": "node ./node_modules/typescript/bin/tsc && npm run bundle",
-    "bundle": "node node_modules/rollup/bin/rollup dist/index.js -f es -o dist/index.js",
+    "build": "node ./node_modules/typescript/bin/tsc",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "publishConfig": {
