+++ packages/renderer-lit-html/src/renderer-lit-html.ts
@@ -1,10 +1,10 @@
-import { render, html } from '../node_modules/lit-html/lib/lit-extended.js';
-export { html } from '../node_modules/lit-html/lib/lit-extended.js';
+import { render, html } from 'lit-html/lib/lit-extended.js';
+export { html } from 'lit-html/lib/lit-extended.js';
 
 export const withLitHtml = (base: any) =>
     class extends base {
         renderer() {
-            if ((<any>this).render()) {
+            if ((<any>this).render) {
                 const root = this.shadowRoot ? this.shadowRoot : this;
 
                 render((<any>this).render(), root);
